move code to src folder. create a db class that maintains a supabase client
